refactor(mirror): split score module into submodules
