+++ apps/mirror/src/score/baseline.rs
@@ -1,0 +1,209 @@
+use chrono::{Duration, Utc};
+
+use super::compute::analyze_tension;
+use super::types::{worst, ScoreResult, Signal};
+
+/// Minimum number of historical scores to activate personal baseline
+const BASELINE_MIN_ENTRIES: usize = 7;
+
+/// Sliding window size in days
+const BASELINE_WINDOW_DAYS: i64 = 28;
+
+/// 4-week rolling averages for each indicator
+#[derive(Debug, Clone)]
+pub struct PersonalBaseline {
+    pub dreyfus_avg: f64,
+    pub decision_quality_avg: f64,
+    pub depth_output_avg: f64,
+    pub exploration_avg: f64,
+    pub deep_invest_avg: f64,
+    pub fragmentation_avg: f64,
+    pub delegation_avg: f64,
+    pub mode_diversity_avg: f64,
+    pub bug_decision_avg: f64,
+    pub knowledge_rate_avg: f64,
+    pub friction_density_avg: f64,
+}
+
+/// Extract a named indicator's actual value from a ScoreResult.
+/// Returns None if the indicator is not found.
+fn extract_indicator(result: &ScoreResult, name: &str) -> Option<f64> {
+    fn canonical_indicator_name(raw: &str) -> &str {
+        match raw {
+            // Legacy localized/label names in historical snapshots.
+            "决策质量" | "Decision Quality" => "decision_quality",
+            "深度产出比" | "Depth Output" => "depth_output",
+            "探索率" | "探索占比" | "Exploration" => "exploration",
+            "深耕率" | "深挖率" | "深挖占比" | "Deep Invest" => "deep_invest",
+            "碎片化" | "Fragmentation" => "fragmentation",
+            "委派率" | "委派比" | "delegation" | "Delegation" => "delegation",
+            "模式多样性" | "Mode Diversity" => "mode_diversity",
+            "bug/决策" | "Bug/Decision" | "bug/decision" => "bug_decision",
+            "知识获取" | "Knowledge" => "knowledge_rate",
+            "摩擦密度" | "Friction" => "friction_density",
+            "Dreyfus" => "dreyfus",
+            _ => raw,
+        }
+    }
+
+    result
+        .layers
+        .iter()
+        .flat_map(|l| &l.indicators)
+        .find(|i| canonical_indicator_name(&i.name) == name)
+        .map(|i| i.actual)
+}
+
+/// Compute personal baseline from historical scores within the last 28 days.
+/// Returns None if fewer than BASELINE_MIN_ENTRIES scores exist in that window.
+pub fn compute_personal_baseline(history: &[ScoreResult]) -> Option<PersonalBaseline> {
+    let cutoff = Utc::now() - Duration::days(BASELINE_WINDOW_DAYS);
+    let recent: Vec<&ScoreResult> = history.iter().filter(|s| s.timestamp >= cutoff).collect();
+
+    if recent.len() < BASELINE_MIN_ENTRIES {
+        return None;
+    }
+
+    let avg = |name: &str| -> f64 {
+        let (sum, count) = recent
+            .iter()
+            .filter_map(|s| extract_indicator(s, name))
+            .fold((0.0, 0usize), |(sum, count), value| {
+                (sum + value, count + 1)
+            });
+
+        if count == 0 {
+            0.0
+        } else {
+            sum / count as f64
+        }
+    };
+
+    Some(PersonalBaseline {
+        dreyfus_avg: avg("dreyfus"),
+        decision_quality_avg: avg("decision_quality"),
+        depth_output_avg: avg("depth_output"),
+        exploration_avg: avg("exploration"),
+        deep_invest_avg: avg("deep_invest"),
+        fragmentation_avg: avg("fragmentation"),
+        delegation_avg: avg("delegation"),
+        mode_diversity_avg: avg("mode_diversity"),
+        bug_decision_avg: avg("bug_decision"),
+        knowledge_rate_avg: avg("knowledge_rate"),
+        friction_density_avg: avg("friction_density"),
+    })
+}
+
+/// Determine signal by comparing actual value against personal baseline.
+/// `higher_is_better`: true for metrics where higher = better (dreyfus, decision_quality, etc.)
+///                     false for metrics where lower = better (delegation, fragmentation, bug_decision)
+pub fn signal_from_personal(actual: f64, baseline: f64, higher_is_better: bool) -> Signal {
+    if baseline == 0.0 {
+        return Signal::Yellow;
+    }
+    let ratio = actual / baseline;
+    if higher_is_better {
+        if ratio >= 1.05 {
+            Signal::Green
+        } else if ratio >= 0.95 {
+            Signal::Yellow
+        } else {
+            Signal::Red
+        }
+    } else {
+        // Lower is better: ratio < 0.95 means actual is notably below baseline = good
+        if ratio <= 0.95 {
+            Signal::Green
+        } else if ratio <= 1.05 {
+            Signal::Yellow
+        } else {
+            Signal::Red
+        }
+    }
+}
+
+/// Indicator config for personal baseline re-judgment
+struct IndicatorMeta {
+    name: &'static str,
+    baseline_value: f64,
+    higher_is_better: bool,
+}
+
+/// Apply personal baseline to override signals on a ScoreResult (in-place).
+pub(super) fn apply_personal_baseline(result: &mut ScoreResult, baseline: &PersonalBaseline) {
+    let metas = [
+        IndicatorMeta {
+            name: "dreyfus",
+            baseline_value: baseline.dreyfus_avg,
+            higher_is_better: true,
+        },
+        IndicatorMeta {
+            name: "decision_quality",
+            baseline_value: baseline.decision_quality_avg,
+            higher_is_better: true,
+        },
+        IndicatorMeta {
+            name: "depth_output",
+            baseline_value: baseline.depth_output_avg,
+            higher_is_better: true,
+        },
+        IndicatorMeta {
+            name: "exploration",
+            baseline_value: baseline.exploration_avg,
+            higher_is_better: true,
+        },
+        IndicatorMeta {
+            name: "deep_invest",
+            baseline_value: baseline.deep_invest_avg,
+            higher_is_better: true,
+        },
+        IndicatorMeta {
+            name: "fragmentation",
+            baseline_value: baseline.fragmentation_avg,
+            higher_is_better: false,
+        },
+        IndicatorMeta {
+            name: "delegation",
+            baseline_value: baseline.delegation_avg,
+            higher_is_better: false,
+        },
+        IndicatorMeta {
+            name: "mode_diversity",
+            baseline_value: baseline.mode_diversity_avg,
+            higher_is_better: true,
+        },
+        IndicatorMeta {
+            name: "bug_decision",
+            baseline_value: baseline.bug_decision_avg,
+            higher_is_better: false,
+        },
+        IndicatorMeta {
+            name: "knowledge_rate",
+            baseline_value: baseline.knowledge_rate_avg,
+            higher_is_better: true,
+        },
+        IndicatorMeta {
+            name: "friction_density",
+            baseline_value: baseline.friction_density_avg,
+            higher_is_better: false,
+        },
+    ];
+
+    for layer in &mut result.layers {
+        for indicator in &mut layer.indicators {
+            if let Some(meta) = metas.iter().find(|m| m.name == indicator.name) {
+                indicator.signal = signal_from_personal(
+                    indicator.actual,
+                    meta.baseline_value,
+                    meta.higher_is_better,
+                );
+            }
+        }
+        // Recalculate layer signal as worst of its indicators
+        let sigs: Vec<Signal> = layer.indicators.iter().map(|i| i.signal).collect();
+        layer.signal = worst(&sigs);
+    }
+
+    // Recalculate tension after signal changes
+    result.tension = analyze_tension(&result.layers);
+}
